# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Model {
  id: string;
  name: string;
  provider: string;
  type: string;
  description?: string;
  trustScore: number;
  accuracy: number;
  safety: number;
  fairness: number;
  hallucination: number;
  robustness: number;
  explainability?: number;
  tests: number;
  lastEvaluated: string;
  latency?: number;
  cost?: number;
}

export interface ModelInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  provider: string;
  /** @minLength 1 */
  type: string;
  description?: string;
}

export type EvaluationStatus = typeof EvaluationStatus[keyof typeof EvaluationStatus];


export const EvaluationStatus = {
  draft: 'draft',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Evaluation {
  id: string;
  name: string;
  modelId: string;
  modelName: string;
  mode: string;
  status: EvaluationStatus;
  trustScore: number;
  tests: number;
  passed: number;
  failed: number;
  critical: number;
  hallucinationRate: number;
  safetyScore: number;
  biasSignals: number;
  createdAt: string;
  updatedAt: string;
  categories?: string[];
}

export type EvaluationInputMode = typeof EvaluationInputMode[keyof typeof EvaluationInputMode];


export const EvaluationInputMode = {
  quick: 'quick',
  standard: 'standard',
  deep: 'deep',
  adversarial: 'adversarial',
} as const;

export interface EvaluationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  modelId: string;
  mode: EvaluationInputMode;
  /** @minItems 1 */
  categories: string[];
  /**
     * @minimum 1
     * @maximum 10000
     */
  tests?: number;
}

export type EvaluationResultSeverity = typeof EvaluationResultSeverity[keyof typeof EvaluationResultSeverity];


export const EvaluationResultSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface EvaluationResult {
  id: string;
  evaluationId: string;
  category: string;
  prompt: string;
  response: string;
  expectedResponse: string;
  score: number;
  confidence: number;
  passed: boolean;
  severity: EvaluationResultSeverity;
  hallucination: string;
  safety: string;
  bias: string;
  reason: string;
  evidence: string;
  recommendation: string;
  createdAt?: string;
}

export type TestSuiteInputDifficulty = typeof TestSuiteInputDifficulty[keyof typeof TestSuiteInputDifficulty];


export const TestSuiteInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface TestSuiteInput {
  category: string;
  difficulty: TestSuiteInputDifficulty;
  /**
     * @minimum 1
     * @maximum 100
     */
  count: number;
}

export type ReportDimensionScores = {[key: string]: number};

export interface Report {
  id: string;
  evaluationId: string;
  title: string;
  status: string;
  trustScore: number;
  modelName: string;
  createdAt: string;
  executiveSummary: string;
  methodology: string;
  recommendations: string[];
  criticalFindings: string[];
  dimensionScores?: ReportDimensionScores;
}

export type ExportInputFormat = typeof ExportInputFormat[keyof typeof ExportInputFormat];


export const ExportInputFormat = {
  json: 'json',
  csv: 'csv',
  pdf: 'pdf',
} as const;

export interface ExportInput {
  format: ExportInputFormat;
}

export interface ExportPayload {
  format: string;
  filename: string;
  content: string;
}

export interface RedTeamInput {
  modelId: string;
  attackTypes?: string[];
}

export interface RedTeamFinding {
  id: string;
  type: string;
  severity: string;
  title: string;
  detail: string;
  recommendation: string;
  status: string;
}

export interface RedTeamReport {
  attackCount: number;
  successfulAttacks: number;
  blockedAttacks: number;
  safetyRate: number;
  criticalVulnerabilities: number;
  findings: RedTeamFinding[];
}

export interface LeaderboardEntry {
  rank: number;
  model: string;
  trustScore: number;
  accuracy: number;
  safety: number;
  fairness: number;
  hallucination: number;
  robustness: number;
  tests: number;
  lastEvaluated: string;
}

export interface DimensionScore {
  name: string;
  score: number;
  fullMark: number;
}

export interface TrendPoint {
  date: string;
  score: number;
}

export interface CountPoint {
  name: string;
  value: number;
}

export interface Dashboard {
  trustScore: number;
  status: string;
  totalTests: number;
  passed: number;
  failed: number;
  critical: number;
  hallucinationRate: number;
  biasRisk: string;
  safetyScore: number;
  dimensions: DimensionScore[];
  trend: TrendPoint[];
  distribution: CountPoint[];
  severityDistribution: CountPoint[];
  recentEvaluations: Evaluation[];
  recentFailures: EvaluationResult[];
}