# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CountPoint } from './countPoint';
import type { DimensionScore } from './dimensionScore';
import type { Evaluation } from './evaluation';
import type { EvaluationResult } from './evaluationResult';
import type { TrendPoint } from './trendPoint';

export interface Dashboard {
  trustScore: number;
  status: string;
  totalTests: number;
  passed: number;
  failed: number;
  critical: number;
  hallucinationRate: number;
  biasRisk: string;
  safetyScore: number;
  dimensions: DimensionScore[];
  trend: TrendPoint[];
  distribution: CountPoint[];
  severityDistribution: CountPoint[];
  recentEvaluations: Evaluation[];
  recentFailures: EvaluationResult[];
}