# lib/api-zod/src/generated/types/dimensionScore.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DimensionScore {
  name: string;
  score: number;
  fullMark: number;
}