# lib/api-zod/src/generated/types/evaluation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationStatus } from './evaluationStatus';

export interface Evaluation {
  id: string;
  name: string;
  modelId: string;
  modelName: string;
  mode: string;
  status: EvaluationStatus;
  trustScore: number;
  tests: number;
  passed: number;
  failed: number;
  critical: number;
  hallucinationRate: number;
  safetyScore: number;
  biasSignals: number;
  createdAt: string;
  updatedAt: string;
  categories?: string[];
}