# lib/api-zod/src/generated/types/evaluationResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationResultSeverity } from './evaluationResultSeverity';

export interface EvaluationResult {
  id: string;
  evaluationId: string;
  category: string;
  prompt: string;
  response: string;
  expectedResponse: string;
  score: number;
  confidence: number;
  passed: boolean;
  severity: EvaluationResultSeverity;
  hallucination: string;
  safety: string;
  bias: string;
  reason: string;
  evidence: string;
  recommendation: string;
  createdAt?: string;
}