# lib/api-zod/src/generated/types/model.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Model {
  id: string;
  name: string;
  provider: string;
  type: string;
  description?: string;
  trustScore: number;
  accuracy: number;
  safety: number;
  fairness: number;
  hallucination: number;
  robustness: number;
  explainability?: number;
  tests: number;
  lastEvaluated: string;
  latency?: number;
  cost?: number;
}