# lib/api-zod/src/generated/types/modelInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  provider: string;
  /** @minLength 1 */
  type: string;
  description?: string;
}