# lib/api-zod/src/generated/types/redTeamFinding.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TrustScore AI evaluation and auditing platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RedTeamFinding {
  id: string;
  type: string;
  severity: string;
  title: string;
  detail: string;
  recommendation: string;
  status: string;
}